refactor: extract useRouterQuery to make useInitSetup testable
